fix(gcp) src/core/zod-helpers.ts. zod-helpers.ts moved to src/core/tools
