+++ src/core/tools/zod-helpers.ts
@@ -4,15 +4,15 @@ import { z } from "zod";
  * Type guard: returns true if `x` is a ZodEnum<[string, ...string[]]>.
  */
 export function isZodEnum(x: unknown): x is z.ZodEnum<[string, ...string[]]> {
-  return x instanceof z.ZodEnum;
+	return x instanceof z.ZodEnum;
 }
 
 /**
  * Type guard: returns true if `x` is a ZodDefault wrapping a ZodEnum.
  */
 export function isZodDefaultEnum(x: unknown): x is z.ZodDefault<z.ZodEnum<[string, ...string[]]>> {
-  return x instanceof z.ZodDefault
-    && ((x as z.ZodDefault<z.ZodTypeAny>)._def?.innerType instanceof z.ZodEnum);
+	return x instanceof z.ZodDefault
+		&& ((x as z.ZodDefault<z.ZodTypeAny>)._def?.innerType instanceof z.ZodEnum);
 }
 
 /**
@@ -21,28 +21,28 @@ export function isZodDefaultEnum(x: unknown): x is z.ZodDefault<z.ZodEnum<[strin
  *  - a ZodDefault<ZodEnum> (by reading its inner enum).
  */
 export function enumOptions<T extends z.ZodEnum<[string, ...string[]]>>(
-  schema: T | z.ZodDefault<T>
+	schema: T | z.ZodDefault<T>
 ): Readonly<T["_def"]["values"]> {
-  if (isZodEnum(schema)) {
-    return schema.options as Readonly<T["_def"]["values"]>;
-  }
-  if (isZodDefaultEnum(schema)) {
-    const inner = (schema as z.ZodDefault<T>)._def.innerType as T;
-    return inner.options as Readonly<T["_def"]["values"]>;
-  }
-  throw new Error("Field is not a ZodEnum or ZodDefault<ZodEnum>.");
+	if (isZodEnum(schema)) {
+		return schema.options as Readonly<T["_def"]["values"]>;
+	}
+	if (isZodDefaultEnum(schema)) {
+		const inner = (schema as z.ZodDefault<T>)._def.innerType as T;
+		return inner.options as Readonly<T["_def"]["values"]>;
+	}
+	throw new Error("Field is not a ZodEnum or ZodDefault<ZodEnum>.");
 }
 
 /**
  * Narrow a loose `string | undefined` to the exact union of literals defined by the schema.
  */
 export function toEnumFromSchema<T extends z.ZodEnum<[string, ...string[]]>>(
-  schema: T | z.ZodDefault<T>,
-  value?: string
+	schema: T | z.ZodDefault<T>,
+	value?: string
 ): T["_def"]["values"][number] | undefined {
-  if (!value) return undefined;
-  const opts = enumOptions(schema) as readonly string[];
-  return opts.includes(value) ? (value as T["_def"]["values"][number]) : undefined;
+	if (!value) return undefined;
+	const opts = enumOptions(schema) as readonly string[];
+	return opts.includes(value) ? (value as T["_def"]["values"][number]) : undefined;
 }
 
 /**
@@ -51,13 +51,13 @@ export function toEnumFromSchema<T extends z.ZodEnum<[string, ...string[]]>>(
  * - Throws an error listing allowed values.
  */
 export function toEnumFromSchemaOrThrow<T extends z.ZodEnum<[string, ...string[]]>>(
-  schema: T | z.ZodDefault<T>,
-  value: string
+	schema: T | z.ZodDefault<T>,
+	value: string
 ): T["_def"]["values"][number] {
-  const v = toEnumFromSchema(schema, value);
-  if (v === undefined) {
-    const opts = Array.from(enumOptions(schema)).join(", ");
-    throw new Error(`Invalid value "${value}". Allowed: ${opts}`);
-  }
-  return v;
+	const v = toEnumFromSchema(schema, value);
+	if (v === undefined) {
+		const opts = Array.from(enumOptions(schema)).join(", ");
+		throw new Error(`Invalid value "${value}". Allowed: ${opts}`);
+	}
+	return v;
 }
